+++ src/solve/LazyGaussianElimination.h
@@ -8,7 +8,7 @@ namespace caramel {
 std::tuple<std::vector<uint32_t>, std::vector<uint32_t>, std::vector<uint32_t>,
            DenseSystemPtr>
 lazyGaussianElimination(const SparseSystemPtr &sparse_system,
-                        const std::vector<uint32_t> &equation_ids);
+                        const std::vector<uint64_t> &equation_ids);
 
 BitArrayPtr solveLazyFromDense(const std::vector<uint32_t> &solved_ids,
                                const std::vector<uint32_t> &solved_vars,
